[UI] Added playback control buttons and fixed minor issues

+++ utils/func/queue.py
@@ -10,11 +10,12 @@ import wavelink
 
 class Queue:
     def __init__(self, info_generator):
-        from ..ui import _sec_to_hms, embed_opt, musicbot, bot
+        from ..ui import _sec_to_hms, embed_opt, musicbot, bot, guild_info
         self._sec_to_hms = _sec_to_hms
         self.embed_opt = embed_opt
         self.musicbot = musicbot
         self.bot = bot
+        self.guild_info = guild_info
         self.info_generator: InfoGenerator = info_generator
 
     # Add to queue
@@ -46,7 +47,7 @@ class Queue:
                 await command.channel.send(msg, embed=embed)
 
     # Queue Embed Generator
-    def _QueueEmbed(self, playlist: PlaylistBase, page: int=0) -> discord.Embed:
+    def _QueueEmbed(self, playlist: PlaylistBase, page: int=0, op=None) -> discord.Embed:
         embed = discord.Embed(title=":information_source: | 候播清單", description=f"以下清單為歌曲候播列表\n共 {len(playlist.order)-1} 首", colour=0xF2F3EE)
         
         for i in range(1, 4):
@@ -62,22 +63,28 @@ class Queue:
         embed_opt = copy.deepcopy(self.embed_opt)
 
         if len(playlist.order) > 4:
+            if op == 'button':
+                txt = '請點按「刪除這些訊息」來關閉\n'
+            else:
+                txt = ''
             total_pages = (len(playlist.order)-1) // 3
             if (len(playlist.order)-1) % 3 != 0:
                 total_pages += 1
-            embed_opt['footer']['text'] = f'第 {page+1} 頁 / 共 {total_pages} 頁\n' + self.embed_opt['footer']['text']
+            embed_opt['footer']['text'] = f'{txt}第 {page+1} 頁 / 共 {total_pages} 頁\n' + self.embed_opt['footer']['text']
         
         embed = discord.Embed.from_dict(dict(**embed.to_dict(), **embed_opt))
         return embed
     
     # Queue Listing
-    async def ShowQueue(self, command: Command) -> None:
+    async def ShowQueue(self, command: Union[Command, discord.Interaction], op=None) -> None:
         playlist: PlaylistBase = self.musicbot._playlist[command.guild.id]
 
         class QueueListing(discord.ui.View):
 
             QueueEmbed = self._QueueEmbed
             embed_opt = self.embed_opt
+            operation = op
+            guild_info = self.guild_info
 
             def __init__(self, *, timeout=60):
                 super().__init__(timeout=timeout)
@@ -123,7 +130,7 @@ class Queue:
             async def firstpage(self, interaction: discord.Interaction, button: discord.ui.Button):
                 self.page = 0
                 self.update_button()
-                embed = self.QueueEmbed(playlist, self.page)
+                embed = self.QueueEmbed(playlist, self.page, self.operation)
                 await interaction.response.edit_message(embed=embed, view=view)
 
             @discord.ui.button(label='⬅️', style=discord.ButtonStyle.gray, disabled=True)
@@ -132,7 +139,7 @@ class Queue:
                 if self.page < 0:
                     self.page = 0
                 self.update_button()
-                embed = self.QueueEmbed(playlist, self.page)
+                embed = self.QueueEmbed(playlist, self.page, self.operation)
                 await interaction.response.edit_message(embed=embed, view=view)
 
             @discord.ui.button(label='➡️', style=discord.ButtonStyle.blurple)
@@ -141,14 +148,14 @@ class Queue:
                 if self.page > self.total_pages:
                     self.page = self.total_pages
                 self.update_button()
-                embed = self.QueueEmbed(playlist, self.page)
+                embed = self.QueueEmbed(playlist, self.page, self.operation)
                 await interaction.response.edit_message(embed=embed, view=view)
 
             @discord.ui.button(label='⏩', style=discord.ButtonStyle.blurple)
             async def lastpage(self, interaction: discord.Interaction, button: discord.ui.Button):            
                 self.page = self.total_pages
                 self.update_button()
-                embed = self.QueueEmbed(playlist, self.page)
+                embed = self.QueueEmbed(playlist, self.page, self.operation)
                 await interaction.response.edit_message(embed=embed, view=view)
 
             @discord.ui.button(label='❎', style=discord.ButtonStyle.danger)
@@ -158,19 +165,37 @@ class Queue:
                 self.stop()
 
             async def on_timeout(self):
-                await msg.delete()
+                if self.operation == 'button':
+                    await command.edit_original_message(content='時限已到，請按「關閉這些訊息」來刪掉此訊息', view=None, embed=None)
+                else:
+                    await msg.delete()
             
         if (len(playlist.order) < 2):
-            await command.send(f'''
+            if op == 'button':
+                await command.response.send_message(f'''
+            **:information_source: | 待播歌曲**
+            目前沒有任何歌曲待播中
+            *輸入 ** '{self.bot.command_prefix}play 關鍵字或網址' **可繼續點歌*
+            *請點按「刪除這些訊息」來關閉此訊息*
+            ''', ephemeral=True)
+            else:
+                await command.send(f'''
             **:information_source: | 待播歌曲**
             目前沒有任何歌曲待播中
             *輸入 ** '{self.bot.command_prefix}play 關鍵字或網址' **可繼續點歌*
             ''')
             return
         else:
-            embed = self._QueueEmbed(playlist, 0)
+            embed = self._QueueEmbed(playlist, 0, op)
             if not (len(playlist.order)) <= 4:
                 view = QueueListing()
-                msg = await command.send(embed=embed, view=view)
+                if op == 'button':
+                    view.remove_item(view.done)
+                    await command.response.send_message(embed=embed, view=view, ephemeral=True)
+                else:
+                    msg = await command.send(embed=embed, view=view)
             else:
-                await command.send(embed=embed)
+                if op == 'button':
+                    await command.response.send_message(embed=embed, ephemeral=True)
+                else:
+                    await command.send(embed=embed)
